backend-app-api: move limited token creation to UserTokenHandler
